+++ src/models/player.lua
@@ -1,3 +1,4 @@
+local anim8         = require("lib.anim8")
 local tween         = require("lib.tween")
 local vector        = require("lib.vector")
 local colors        = require("src.globals.colors")
@@ -9,16 +10,20 @@ local player        = {}
 -- === Constants ===
 local FLIP_DURATION = 0.5
 
-function player:init()
-    self.w, self.h  = 32, 16
+function player:init(tileset)
+    self.w, self.h  = 48, 48
     self.speed      = 400
     self.lane       = 1 -- 1=left, 2=right
     self.isFlipping = false
     self.flipTimer  = FLIP_DURATION
     self.pos        = vector(
-        consts.LANE_WIDTH - self.w / 2,
+        consts.LANE_WIDTH,
         consts.WINDOW_HEIGHT * 0.75
     )
+
+    -- Animation
+    local grid      = anim8.newGrid(16, 16, tileset:getWidth(), tileset:getHeight())
+    self.animation  = anim8.newAnimation(grid("1-4", 1), 0.1)
 end
 
 -- === Behavior ===
@@ -37,6 +42,8 @@ function player:update(dt)
             self.flipTimer = 1
             self.lane = self.lane == 1 and 2 or 1
         end
+    else
+        self.animation:update(dt)
     end
 end
 
@@ -47,9 +54,9 @@ function player:jump()
 
         local newX
         if self.lane == 1 then
-            newX = consts.WINDOW_WIDTH - consts.LANE_WIDTH - self.w * 1.5
+            newX = consts.WINDOW_WIDTH - consts.LANE_WIDTH - self.w
         else
-            newX = consts.LANE_WIDTH - self.w / 2
+            newX = consts.LANE_WIDTH
         end
 
         self.posTween = tween.new(FLIP_DURATION, self.pos, { x = newX }, "outQuart")
@@ -67,11 +74,12 @@ function player:checkCollision(obstacles)
     return hit
 end
 
-function player:draw()
+function player:draw(tileset)
     love.graphics.setColor(colors.SLATE_800)
 
-    local origin = self.pos + vector(self.w / 2, self.h / 2)
-    love.graphics.rectangle("fill", origin.x, origin.y, self.w, self.h)
+    local oX, oY = self.w / 2, self.h / 2
+    local rotation = self.lane == 1 and math.pi / 2 or -math.pi / 2
+    self.animation:draw(tileset, self.pos.x + oX, self.pos.y + oY, rotation, 3, 3, 8, 8)
 end
 
 return player